moved Cities, distance formula, permutation to an import file
