+++ exhaustiveSalesman.js
@@ -1,38 +1,10 @@
 
-// import perm from './minilab/permutation';
+const importfile = require('./minilab/importfile');
+const perm = importfile.perm;
+const Cities = importfile.Cities;
 
-const perm = (args) => {
-  const result = [];
-  const permute = (arr, memo = []) => {
-    let i, current;
-    for (i = 0; i < arr.length; i++) {
-      current = arr.splice(i, 1);
-      if (arr.length === 0) {
-        result.push(memo.concat(current));
-      }
-      permute(arr.slice(), memo.concat(current));
-      arr.splice(i, 0, current[0]);
-    }
-    return result;
-  };
-  return permute(args);
-};
-
-
-distance_between = (city1, city2) => {
-  return Math.sqrt(Math.pow((city1.x - city2.x), 2) + Math.pow((city1.y - city2.y), 2));
-}
-
-
-const Cities = [
-  {name:"Salt Lake City", x:300, y:500},
-  {name:"Santa Fe", x:500, y:350},
-  {name:"Denver", x:500, y:500},
-  {name:"Cheyenne", x:500, y:600},
-  // {name: "New York", x:1000, y: 1000},
-];
-
-let BTL = 999999999999;
+let BTL = Number.MAX_SAFE_INTEGER;
+// console.log(BTL);
 let BT = [];
 
 // make the permutations based on how many cities there are
@@ -61,3 +33,5 @@ for(let i = 0; i < permCities.length; i++) {
     }
   }
 }
+// console.log(BTL);
+// console.log(BT);
